Update build configuration and paths
In this commit, the build configuration and paths have been updated to
simplify the build process. The usage of `tsconfig.esm.json` has been
removed in favor of a single `tsconfig.json` file. The
`mainEntryPointFilePath` in `api-extractor.json` and the `build:ts`
script in `package.json` have also been modified accordingly.