Add error handling for unused variables
Implemented function 'findExtraVariables' to identify any variables
provided by the user that are not used in the prompt. The function then
throws an error if any unused variables are found, informing the user
about the unused variables. This change improves error handling by
notifying the user of potential mistakes in their provided input.